I12  ownership

+++ core/src/api/connections.ts
@@ -1,10 +1,11 @@
-import { Router, Request, Response } from 'express'
+import { Router, Response, request } from 'express'
 import { ObjectId } from 'mongodb'
 import { ConnectionStringParser } from 'connection-string-parser'
 import ConnectionsFactory from '../connections'
 import MongoConnection from '../connections/mongo'
 import { APIError, withCatch } from './utils'
 import { encrypt } from '../crypto'
+import { AuthRequest, requiresAuth } from '../middlewares'
 
 export enum SUPPORTED_DRIVERS {
   'mongodb' = 'mongodb',
@@ -22,6 +23,9 @@ export type BIDataConnectionParams =
   | BIDataPostgresConnectionParams
 export type BIDataConnection = {
   _id: ObjectId
+  owner_id: ObjectId
+  creator_id: ObjectId
+  private: boolean
   name: string
   driver: SUPPORTED_DRIVERS
   params: BIDataConnectionParams
@@ -58,24 +62,42 @@ const validateConnectionData = (connection: BIDataConnection) => {
   }
 }
 
-const handleListConnections = async (req: Request, res: Response) => {
+const handleListConnections = async (req: AuthRequest, res: Response) => {
   const core = (await ConnectionsFactory.get()) as MongoConnection
-  const connections = await core.client.collection('Connections').find()
+  const connections = await core.client.collection('Connections').find({
+    $or: [
+      {
+        creator_id: req.user._id,
+      },
+      {
+        private: false,
+      },
+    ],
+  })
   const connectionsAsArr = await connections.toArray()
   return res.send(connectionsAsArr)
 }
 
-const handleGetConnection = async (req: Request, res: Response) => {
+const handleGetConnection = async (req: AuthRequest, res: Response) => {
   const id = new ObjectId(req.params.id)
   const core = (await ConnectionsFactory.get()) as MongoConnection
   const connection = await core.client
     .collection('Connections')
     .findOne({ _id: id })
 
+  if (!connection) {
+    throw new APIError('Connection not found', 404)
+  }
+  if (connection.private && !connection.owner_id.equals(req.user._id)) {
+    throw new APIError(
+      'You do not have permissions to see this connection',
+      401
+    )
+  }
   return res.send(connection)
 }
 
-const handleCreateConnection = async (req: Request, res: Response) => {
+const handleCreateConnection = async (req: AuthRequest, res: Response) => {
   const body = {
     ...req.body,
   } as BIDataConnection
@@ -124,14 +146,20 @@ const handleCreateConnection = async (req: Request, res: Response) => {
 
   const connectionDocument = await core.client
     .collection('Connections')
-    .insertOne(body)
+    .insertOne({
+      ...body,
+      creator_id: req.user._id,
+      owner_id: req.user._id,
+      created: new Date(),
+      private: true,
+    })
   return res.send({
     _id: connectionDocument.insertedId,
     ...body,
   })
 }
 
-const handleUpdateConnection = async (req: Request, res: Response) => {
+const handleUpdateConnection = async (req: AuthRequest, res: Response) => {
   const id = new ObjectId(req.params.id)
   const body = req.body as BIDataConnection & { _id?: string }
   delete body._id
@@ -143,35 +171,66 @@ const handleUpdateConnection = async (req: Request, res: Response) => {
     return res.status(e.status).send(e)
   }
 
-  try {
-    const connection = ConnectionsFactory.createConnection(body)
-    await connection.checkConectivity()
-  } catch (e) {
-    const apiError = new APIError(e.message, 500)
-    return res.status(apiError.status).send(apiError)
-  }
+  const connection = ConnectionsFactory.createConnection(body)
+  await connection.checkConectivity()
 
   const core = ConnectionsFactory.get() as MongoConnection
+
+  const existingConnection = await core.client
+    .collection('Connections')
+    .findOne({ _id: id })
+
+  if (!existingConnection) {
+    throw new APIError('Connection does not exists', 404)
+  }
+
+  if (!existingConnection.owner_id.equals(req.user._id)) {
+    throw new APIError('Do not have rights to edit the connection', 401)
+  }
 
   const updatedConnection = await core.client
     .collection('Connections')
-    .findOneAndUpdate({ _id: id }, { $set: body }, { returnOriginal: false })
+    .findOneAndUpdate(
+      { _id: id },
+      {
+        $set: {
+          ...body,
+          updated: new Date(),
+        },
+      },
+      { returnOriginal: false }
+    )
   return res.send(updatedConnection.value)
 }
 
-const handleDeleteConnection = async (req: Request, res: Response) => {
+const handleDeleteConnection = async (req: AuthRequest, res: Response) => {
   const id = new ObjectId(req.params.id)
   const core = ConnectionsFactory.get() as MongoConnection
+
+  const existingConnection = await core.client
+    .collection('Connections')
+    .findOne({ _id: id })
+
+  if (!existingConnection) {
+    throw new APIError('Connection does not exists', 404)
+  }
+
+  if (!existingConnection.owner_id.equals(req.user._id)) {
+    throw new APIError('Do not have rights to edit the connection', 401)
+  }
 
   await core.client.collection('Connections').deleteOne({ _id: id })
   return res.status(200).send()
 }
 
-const handleGetDrivers = async (req: Request, res: Response) => {
+const handleGetDrivers = async (req: AuthRequest, res: Response) => {
   return res.send(Object.keys(SUPPORTED_DRIVERS))
 }
 
-const handleGetConnectionConnectivity = async (req: Request, res: Response) => {
+const handleGetConnectionConnectivity = async (
+  req: AuthRequest,
+  res: Response
+) => {
   const id = new ObjectId(req.params.id)
   const core = ConnectionsFactory.get() as MongoConnection
   const connectionDesction = await core.client
@@ -186,11 +245,15 @@ const handleGetConnectionConnectivity = async (req: Request, res: Response) => {
   return res.send()
 }
 
-ConnectionsAPI.get('/', withCatch(handleListConnections))
-ConnectionsAPI.get('/drivers', withCatch(handleGetDrivers))
-ConnectionsAPI.get('/:id', withCatch(handleGetConnection))
-ConnectionsAPI.get('/:id/connectivity', withCatch(handleGetConnectionConnectivity))
-ConnectionsAPI.post('/', withCatch(handleCreateConnection))
-ConnectionsAPI.put('/:id', withCatch(handleUpdateConnection))
-ConnectionsAPI.delete('/:id', withCatch(handleDeleteConnection))
+ConnectionsAPI.get('/', requiresAuth(), withCatch(handleListConnections))
+ConnectionsAPI.get('/drivers', requiresAuth(), withCatch(handleGetDrivers))
+ConnectionsAPI.get('/:id', requiresAuth(), withCatch(handleGetConnection))
+ConnectionsAPI.get(
+  '/:id/connectivity',
+  requiresAuth(),
+  withCatch(handleGetConnectionConnectivity)
+)
+ConnectionsAPI.post('/', requiresAuth(), withCatch(handleCreateConnection))
+ConnectionsAPI.put('/:id', requiresAuth(), withCatch(handleUpdateConnection))
+ConnectionsAPI.delete('/:id', requiresAuth(), withCatch(handleDeleteConnection))
 export default ConnectionsAPI
